add createProduct method to ProductController; implement product creation with category validation

+++ src/main/java/sn/sdley/springbootstarter0/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package sn.sdley.springbootstarter0.repositories;
 import org.springframework.data.repository.CrudRepository;
 import sn.sdley.springbootstarter0.entities.Category;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Byte> {
 }
